feat: share agent runner
